extracted address and email fields
